chore(gap): move gap setters to box renderable
