+++ packages/core/src/renderables/Text.ts
@@ -7,7 +7,7 @@ import { type SelectionState, type RenderContext } from "../types"
 import type { OptimizedBuffer } from "../buffer"
 import { MeasureMode } from "yoga-layout"
 
-export interface TextOptions extends Omit<RenderableOptions, "gap" | "rowGap" | "columnGap"> {
+export interface TextOptions extends RenderableOptions {
   content?: StyledText | string
   fg?: string | RGBA
   bg?: string | RGBA
